+++ packages/assemble-release-plan/src/utils.ts
@@ -2,6 +2,7 @@ import type { PackageGroup, VersionType } from "@changesets/types";
 import type { Package } from "@manypkg/get-packages";
 import semverGt from "semver/functions/gt.js";
 import type { InternalRelease } from "./types.ts";
+import { InternalError } from "@changesets/errors";
 
 export function getHighestReleaseType(
   releases: InternalRelease[]
@@ -39,14 +40,11 @@ export function getCurrentHighestVersion(
   let highestVersion: string | undefined;
 
   for (let pkgName of packageGroup) {
-    let pkg = packagesByName.get(pkgName);
-
-    if (!pkg) {
-      console.error(
-        `FATAL ERROR IN CHANGESETS! We were unable to version for package group: ${pkgName} in package group: ${packageGroup.toString()}`
-      );
-      throw new Error(`fatal: could not resolve linked packages`);
-    }
+    let pkg = mapGetOrThrowInternal(
+      packagesByName,
+      pkgName,
+      `We were unable to version for package group: ${pkgName} in package group: ${packageGroup.toString()}`
+    );
 
     if (
       highestVersion === undefined ||
@@ -57,4 +55,28 @@ export function getCurrentHighestVersion(
   }
 
   return highestVersion!;
+}
+
+export function mapGetOrThrow<V extends {}>(
+  map: Map<string, V>,
+  key: string,
+  errorMessage: string
+): V {
+  const value = map.get(key);
+  if (value === undefined) {
+    throw new Error(errorMessage);
+  }
+  return value;
+}
+
+export function mapGetOrThrowInternal<V extends {}>(
+  map: Map<string, V>,
+  key: string,
+  errorMessage: string
+): V {
+  const value = map.get(key);
+  if (value === undefined) {
+    throw new InternalError(errorMessage);
+  }
+  return value;
 }